requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
bitstring=4.0.1=pypi_0
blas=1.0=mkl
blessings=1.7=py37h06a4308_1002
bottleneck=1.3.5=py37h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=py37h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=38.0.4=py37h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py37h295c915_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.6=py37h06a4308_0
entrypoints=0.4=pyhd8ed1ab_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
fftw=3.3.9=h27cfd23_1
flit-core=3.6.0=pyhd3eb1b0_0
fontconfig=2.14.1=h52c9d5c_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gpustat=0.6.0=pyhd3eb1b0_1
grpcio=1.51.1=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.4=py37h06a4308_0
importlib-metadata=6.0.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.15.0=pyh210e3f2_0
ipython=7.33.0=py37h89c1867_0
jedi=0.18.2=pyhd8ed1ab_0
joblib=1.1.1=py37h06a4308_0
jpeg=9e=h7f8727e_0
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=4.11.1=py37h06a4308_0
kiwisolver=1.4.4=py37h6a678d5_0
krb5=1.19.4=h568e23c_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libdeflate=1.8=h7f8727e_5
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=hecacb30_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lz4-c=1.9.4=h6a678d5_0
matplotlib=3.5.2=py37h06a4308_0
matplotlib-base=3.5.2=py37hf590b9c_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
munkres=1.1.4=py_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.4=py37he184ba9_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
nvidia-ml=7.352.0=pyhd3eb1b0_0
opencv-python=4.7.0.68=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
packaging=22.0=py37h06a4308_0
pandas=1.3.5=py37h8c16a72_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.3.0=py37hace64e9_1
pip=22.3.1=py37h06a4308_0
ply=3.11=py37_0
prettytable=3.6.0=pypi_0
prompt-toolkit=3.0.36=pyha770c72_0
psutil=5.9.0=py37h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
py3nvml=0.2.7=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.14.0=pyhd8ed1ab_0
pynvml=11.4.1=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py37h06a4308_0
pyqt=5.15.7=py37h6a678d5_1
pyqt5-sip=12.11.0=py37h6a678d5_1
pysocks=1.7.1=py37_1
python=3.7.16=h7a1cb2a_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.12.1=py3.7_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2022.7=py37h06a4308_0
pyzmq=19.0.2=py37hac76be4_2
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.28.1=py37h06a4308_0
robustness=9=dev_0
scikit-learn=1.0.2=py37h51133e4_1
scipy=1.7.3=py37h6c91a56_2
seaborn=0.12.2=py37h06a4308_0
setuptools=65.6.3=py37h06a4308_0
sip=6.6.2=py37h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0.post1=pypi_0
smmap=5.0.0=pypi_0
sqlite=3.40.1=h5082296_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.12.1=py37_cu113
torchvision=0.13.1=py37_cu113
tornado=6.2=py37h5eee18b_0
tqdm=4.64.1=py37h06a4308_0
traitlets=5.9.0=pyhd8ed1ab_0
typing_extensions=4.4.0=py37h06a4308_0
urllib3=1.26.14=py37h06a4308_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
xlwt=1.3.0=pypi_0
xmltodict=0.13.0=pypi_0
xz=5.2.10=h5eee18b_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.11.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
